# swagger.yaml
openapi: 3.0.1
info:
  title: Get Science API
  description: API платформы для научных и образовательных мероприятий
  version: 1.0.0
  contact:
    name: Get Science Team
    url: https://get-science-events.ru
servers:
  - url: https://get-science-events.ru/api/v1
    description: Основной сервер
...
paths:
...
  /address/suggest:
    get:
      tags:
        - address
      summary: Поиск подсказок адресов
      description: Возвращает подсказки адресов по заданному поисковому запросу
      operationId: suggest
      parameters:
        - name: query
          in: query
          description: Поисковый запрос
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Ограничение количества результатов
          required: false
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Список подсказок успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressSuggestionDTO'
        '400':
          description: Пустой поисковый запрос
...
components:
  schemas:
...
    AddressSuggestionDTO:
      type: object
      properties:
        label:
          type: string
          description: Полный адрес
        latitude:
          type: number
          format: double
          description: Широта
        longitude:
          type: number
          format: double
          description: Долгота
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []